refactor: 쿼리 파라미터 사용
