+++ src/hooks/useOrderForm.ts
@@ -1,0 +1,6 @@
+import { useState } from 'react';
+
+export const useOrderForm = () => {
+  const [isFilled, setIsFilled] = useState('');
+  const [phoneNumber, setPhoneNumber] = useState('');
+};
